Refactor rank and filter functionality to rank.py Added helpers/common.py for shared code between other modules.
